requirements.txt
numpy==1.26.3
pyyaml==6.0.1